+++ test/services/imageCreatorSpec.scala
@@ -10,17 +10,17 @@ import scala.collection.mutable.ListBuffer
 class imageCreatorSpec extends AnyFunSuite{
     val arrayToImgTest = new ArrayToImage
     test("Create blank buffered image"){
-        val testArray = Array.ofDim[Int](100,100)
+        val testArray = Array.ofDim[Byte](100,100)
         assert(arrayToImgTest.createImg(testArray, 13158600, 20).isInstanceOf[BufferedImage])
     }
 
     test("Test image height dimension"){
-        val testArray = Array.ofDim[Int](100,150)
+        val testArray = Array.ofDim[Byte](100,150)
         assert(arrayToImgTest.createImg(testArray, 13158600, 20).getHeight == testArray.length*20)
     }
 
     test("Test image width dimension"){
-        val testArray = Array.ofDim[Int](100,150)
+        val testArray = Array.ofDim[Byte](100,150)
         assert(arrayToImgTest.createImg(testArray, 13158600, 20).getWidth == testArray(0).length*20)
     }
 }
